refactor: move merge output filename logic to agent_reflect module
