+++ coauthor/__init__.py
@@ -1,4 +1,5 @@
 from .agent_reflect import ThinkAndWrite, DirectWrite
+from .agent_reflect import get_output_file_name_merge
 from .arg_utils import get_common_argparser, comma_separated_list
 from .figure_tools import extract_figure_paths, extract_and_compile_tikzpictures_with_labels
 from .file_utils import (
