+++ coauthor/config.py
@@ -104,6 +104,10 @@ class TaskConfig:
     def from_args(cls, args) -> "TaskConfig":
         """Create a TaskConfig from command line arguments."""
         config = cls(
+            # Processing configuration
+            reflect=args.reflect,
+            model=args.model,
+            agent=args.agent,
             # Input/Output configuration
             input_file=args.input_file,
             input_files=args.input_files,
@@ -117,12 +121,8 @@ class TaskConfig:
             output_files=args.output_files,
             output_name_override=args.output_name_override,
             instruction=args.instruction,
-            # Processing configuration
-            reflect=args.reflect,
-            use_prefill_from_input=args.use_prefill_from_input,
-            model=args.model,
-            agent=args.agent,
             # Tool usage configuration
+            use_prefill_from_input=args.use_prefill_from_input,
             include_tex_count=args.include_tex_count,
             auto_extract_figure=args.auto_extract_figure,
             auto_extract_tikz_figure=args.auto_extract_tikz_figure,
